Extract toChange to adapter-utils and use it all around the code instead of duplicating the logic per each test utils
