+++ packages/adapter-api/index.ts
@@ -23,4 +23,3 @@ export * from './src/change'
 export * from './src/change_group'
 export * from './src/utils'
 export * from './src/error'
-export * from './test/utils'
