+++ packages/core/test/core/plan/plan_item.test.ts
@@ -16,9 +16,9 @@
 import wu from 'wu'
 import _ from 'lodash'
 import { Group } from '@salto-io/dag'
-import { isListType, Change, ChangeDataType } from '@salto-io/adapter-api'
+import { isListType, Change, ChangeDataType, toChange } from '@salto-io/adapter-api'
 import * as mock from '../../common/elements'
-import { getFirstPlanItem, toChange } from '../../common/plan'
+import { getFirstPlanItem } from '../../common/plan'
 import { planGenerators } from './plan.test'
 import { PlanItem } from '../../../src/core/plan'
 import { addPlanItemAccessors } from '../../../src/core/plan/plan_item'
